Perform a consistent set of checks across core-services to ensure that the limit (MaxResultCount) is not exceeded.
