+++ internal/support/logging/utils.go
@@ -1,0 +1,8 @@
+package logging
+
+func checkMaxLimitCount(limit int) int {
+	if limit > Configuration.Service.MaxResultCount || limit == 0 {
+		return Configuration.Service.MaxResultCount
+	}
+	return limit
+}
